refactor: remove dead native nodestore backend path (#768)
storage/nodestore carried two implementations of its Database interface.
Only the kvstore-backed wrapper (KVDatabaseImpl) is wired in production
(internal/cli/server.go); the original native-backend path was superseded
when the kvstore abstraction was introduced but never deleted.

Remove the dead path and everything that only it reached:
- database.go (DatabaseImpl/NewDatabase), backend.go (Backend factory),
  memory.go (MemoryBackend), pebble.go (PebbleBackend)
- verify.go (BackendVerifier) — wraps the Backend interface, so it cannot
  outlive it; the live wrapper exposes no verifier
- config.go in full (its Config type fed only the backend factory; the
  live DatabaseConfig lived in database.go, not config.go as the issue
  surmised — it is relocated into kvstore_database.go alongside its only
  consumer, with newAsyncSem/asyncWorkerLimit)
- types.go: Backend interface, Status, Hash256FromData (no remaining caller)
- errors.go: the Is* helpers (no remaining caller); exported Err* sentinels kept
- the backend-only tests (memory/verify/database_extended/bench) plus the
  backend half of nodestore_test.go; TestNode/TestBatches stay as the sole
  coverage of the kept Node/hash helpers

The persistence.go doc comment is repointed from DatabaseImpl.Sync to
KVDatabaseImpl.Sync. `deadcode -test ./...` reports the nodestore package
clean after the change; `just build` and `just test-libs` pass.

Fixes #768
